pyproject.toml
[tool.poetry]
name = "github-stat-to postgres"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "github_stat_to postgres"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
